# ResolveFieldContext.Path optimization