Slight performance enhancement

```
BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18362
Intel Core i7-7740X CPU 4.30GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.1.100
  [Host]     : .NET Core 3.1.0 (CoreCLR 4.700.19.56402, CoreFX 4.700.19.56404), X64 RyuJIT
  DefaultJob : .NET Core 3.1.0 (CoreCLR 4.700.19.56402, CoreFX 4.700.19.56404), X64 RyuJIT

** before **

|        Method |        Mean |     Error |    StdDev |    Gen 0 |   Gen 1 | Gen 2 |  Allocated |
|-------------- |------------:|----------:|----------:|---------:|--------:|------:|-----------:|
| Introspection | 1,882.08 us | 23.988 us | 21.265 us | 255.8594 | 99.6094 |     - | 1318.82 KB |
|          Hero |    44.97 us |  0.443 us |  0.414 us |   7.4463 |       - |     - |   30.76 KB |

** after **

|        Method |        Mean |     Error |    StdDev |    Gen 0 |   Gen 1 | Gen 2 |  Allocated |
|-------------- |------------:|----------:|----------:|---------:|--------:|------:|-----------:|
| Introspection | 1,818.70 us | 14.744 us | 13.070 us | 259.7656 | 99.6094 |     - | 1319.12 KB |
|          Hero |    44.59 us |  0.383 us |  0.359 us |   7.5073 |       - |     - |   30.86 KB |
```