@Shane32 2020-01-16
Relates to #1147