@Shane32 2020-01-16
If these optimizations are already in #1456 then just close this :)