@Shane32 2020-01-16
I actually think it will complement the int to string optimization in #1456