@Shane32 2020-01-16
Note that this could likely be optimized further, by walking up the tree as described in #1147