@sungam3r 2020-01-16
I think it’s not worth changing it now. The difference is too small. In addition, I was going to change this place in a complete different way. Let's leave PR open for now to be able to refer during the discussion.